chore: rename NewTable components to Table
